Add staging site sync monitor component to site dashboard
